// pos-server/src/client/order/dto/order-item.dto.ts
export class OrderItemDto {

    orderNo: string;
    
    productId: number;

    productCode: string;

    productName: string;

    categoryCode: string;
    categoryName: string;
    brandCode: string;
    brandName: string;
        
    unitId: number;

    unitName: string;

    childUnitId: number;

    childUnitName: string;

    unitCost: number;

    cost: number;

    sellPrice: number;
 
    realSellPrice: number;

    qty: number;

    price: number;

    realPrice: number;

}